Make JarJar handle jars with misplaced classes more reasonably.
A "misplaced class" here refers to a class whose jar entry path
(eg, "org/pantsbuild/Foobar.class") does not match the path
implied by its fully-qualified classname.

Previously, if you had a (corrupt) jar with a directory containing
misplaced classes, jarjar would move all the classes in the so
that their directory properly matched their package names. This
was simply an undocumented (and probably unconsidered) side-effect
of how jar entries are transformed by the JarTransformer.

This would cause problems when there already were classes at those
locations, and JarJar would explode due to a duplicate jar entries
problem that it had constructed itself.

With this change, JarJar allows the specification of a
"misnamedClassStrategy" variable, which can be "fatal" (for
fail-fast exit behavior), "move" (for the previously existing
behavior), "skip" (for leaving the class where it is), or "omit"
(removing the class from the output jar).

This defaults to "omit". A warning is also printed to stderr if the
verbose flag is turned on.