Update API version in visitors to ASM5
We are using "org.pantsbuild" % "jarjar" % "1.6.0" in the SBT
build for Scala.

We see an error when we call JarJar with bytecode that contains
the `MethodParameters` attribute.

This seems to be due to the fact that you updated to ASM5, but
did not update all visitors in this project to the ASM5 api
version.

The ASM5's `ClassReader` visits these parameters, ultimately
with `EmptyClassVisitor`, which runs into:

```
    public void visitParameter(String name, int access) {
        if (api < Opcodes.ASM5) {
            throw new RuntimeException();
        }
        if (mv != null) {
            mv.visitParameter(name, access);
        }
    }
```

See:

  https://docs.oracle.com/javase/tutorial/reflect/member/methodparameterreflection.html

for discussion of the `MethodParameters` classfile attribute. It is
not enabled by default, one needs to enable it with:

```
javac -help 2>&1 | grep parameters
  -parameters                Generate metadata for reflection on method parameters
```

No test is included here as I'm not familiar with pants to know how to
add this.